EasyTest

Documentation

data Status

Constructors

Failed

Passed !Int

Skipped

Pending

combineStatus :: Status -> Status -> Status

data Env

Constructors

Env

Fields

rng :: TVar StdGen

messages :: String

results :: TBQueue (Maybe (TMVar (String, Status)))

note_ :: String -> IO ()

allow :: String

newtype Test a

Constructors

Test (ReaderT Env IO (Maybe a))

io :: IO a -> Test a

atomicLogger :: IO (String -> IO ())

expect' :: HasCallStack => Bool -> Test ()

expect :: HasCallStack => Bool -> Test ()

expectEqual' :: (HasCallStack, Eq a, Show a) => a -> a -> Test ()

expectEqual :: (HasCallStack, Eq a, Show a) => a -> a -> Test ()

expectNotEqual :: (HasCallStack, Eq a, Show a) => a -> a -> Test ()

expectJust :: HasCallStack => Maybe a -> Test a

expectRight :: HasCallStack => Either e a -> Test a

expectLeft :: HasCallStack => Either e a -> Test e

tests :: [Test ()] -> Test ()

runOnly :: String -> Test a -> IO ()

Run all tests whose scope starts with the given prefix

rerunOnly :: Int -> String -> Test a -> IO ()

Run all tests with the given seed and whose scope starts with the given prefix

run :: Test a -> IO ()

rerun :: Int -> Test a -> IO ()

run' :: Int -> (String -> IO ()) -> String -> Test a -> IO ()

scope :: String -> Test a -> Test a

Label a test. Can be nested. A `.` is placed between nested scopes, so `scope "foo" . scope "bar"` is equivalent to `scope "foo.bar"`

note :: String -> Test ()

Log a message

note' :: Show s => s -> Test ()

Log a showable value

random :: Random a => Test a

Generate a random value

random' :: Random a => a -> a -> Test a

Generate a bounded random value. Inclusive on both sides.

bool :: Test Bool

word8 :: Test Word8

char :: Test Char

Generate a random Char

int :: Test Int

Generate a random Int

double :: Test Double

Generate a random Double

word :: Test Word

Generate a random Word

int' :: Int -> Int -> Test Int

Generate a random Int in the given range Note: int 0 5` includes both `0` and `5`

char' :: Char -> Char -> Test Char

Generate a random Char in the given range Note: char a z' includes both `a' and `z'.

double' :: Double -> Double -> Test Double

Generate a random Double in the given range Note: double 0 1` includes both `0` and `1`.

word' :: Word -> Word -> Test Word

Generate a random Double in the given range Note: word 0 10` includes both `0` and `10`.

word8' :: Word8 -> Word8 -> Test Word8

Generate a random Double in the given range Note: word8 0 10` includes both `0` and `10`.

pick :: [a] -> Test a

Sample uniformly from the given list of possibilities

picker :: Int -> [a] -> Int -> Maybe a

listOf :: Int -> Test a -> Test [a]

Alias for replicateM

listsOf :: [Int] -> Test a -> Test [[a]]

Generate a list of lists of the given sizes, an alias for `sizes forM n -> listOf n gen`

pair :: Test a -> Test b -> Test (a, b)

Alias for `liftA2 (,)`.

tuple2 :: (Random a, Random b) => Test (a, b)

Alias for pair.

tuple3 :: (Random a, Random b, Random c) => Test (a, b, c)

Generate a random 3-tuple.

tuple4 :: (Random a, Random b, Random c, Random d) => Test (a, b, c, d)

Generate a random 4-tuple.

mapOf :: Ord k => Int -> Test k -> Test v -> Test (Map k v)

Generate a `Data.Map k v` of the given size.

mapsOf :: Ord k => [Int] -> Test k -> Test v -> Test [Map k v]

Generate a `[Data.Map k v]` of the given sizes.

wrap :: Test a -> Test a

Catch all exceptions that could occur in the given Test

runWrap :: Env -> ReaderT Env IO (Maybe a) -> IO (Maybe a)

using :: IO r -> (r -> IO ()) -> (r -> Test a) -> Test a

A test with a setup and teardown

currentScope :: Test String

The current scope

noteScoped :: String -> Test ()

Prepend the current scope to a logging message

ok :: Test ()

Record a successful test at the current scope

done :: Test a

Skip any tests depending on the return value.

skip :: Test ()

Explicitly skip this test

crash :: HasCallStack => String -> Test a

Record a failure at the current scope

nologging :: HasCallStack => Test a -> Test a

Overwrites the env so that note_ (the logger) is a no op

attempt :: Test a -> Test (Maybe a)

Run a test under a new scope, without logs and suppressing all output

pending :: HasCallStack => Test a -> Test a

Placeholder wrapper for a failing test. The test being wrapped is expected/known to fail. Will produce a failure if the test being wrapped suddenly becomes a success.

putResult :: Status -> ReaderT Env IO ()

fork :: Test a -> Test ()

Run a test in a separate thread, not blocking for its result.

fork' :: Test a -> Test (Test a)

Run a test in a separate thread, return a future which can be used to block on its result.